% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsR.R
\name{addMissingValues}
\alias{addMissingValues}
\title{Function for replacing values with missing NA value for a given dataframe and variable}
\usage{
addMissingValues(df_name, col_name, pc = 10)
}
\arguments{
\item{col_name}{Name of the variable}

\item{df}{Name of the dataframe}

\item{pct}{Percent of missing values}
}
\value{
dataframe
}
\description{
Function for replacing values with missing NA value for a given dataframe and variable
}
\examples{
addMissingValues(df=df, col_name=var, pc=10);

}
