% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsR.R
\name{dsR}
\alias{dsR}
\title{Function for creating dataset with given random values for specified variables}
\usage{
dsR(vr = "ms:3;bi:4;ii:1", nr = 100)
}
\arguments{
\item{vr}{Array of variables with types for creating a data frame.}

\item{nr}{Number of rows for given dataframe!}
}
\value{
dataframe
}
\description{
Function for creating dataset with given random values for specified variables
}
\examples{
dsR(vr="ms:3;bi:4;ii:1", nr=100);

}
