% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsR.R
\name{getType}
\alias{getType}
\title{Creating the datasetR dataframe}
\usage{
getType(var_enum_enter = 1)
}
\arguments{
\item{var_enum_enter}{The value enumerator for given vector of values}
}
\value{
The data type for given variable
}
\description{
Get the type of the variable in the vector for given position
}
\examples{
getType(1);
getType(3);

}
