% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsR.R
\name{getValues}
\alias{getValues}
\title{Get the starting set of values for constructing the dataframe}
\usage{
getValues()
}
\value{
This function returns a data frame.
}
\description{
Get the starting set of values for constructing the dataframe
}
\examples{
getValues()

}
