% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsR.R
\docType{data}
\name{set_of_val}
\alias{set_of_val}
\title{Sample data}
\format{
An object of class \code{list} of length 21.
}
\usage{
set_of_val
}
\description{
Sample data
}
\examples{
dsR(vr="ms:3;bi:4;ii:1", nr=100);

}
\keyword{datasets}
